use f64

+++ src/color.rs
@@ -3,7 +3,7 @@ use crate::vec3::Vec3;
 pub type Color = Vec3;
 
 impl Color {
-    pub fn from_tup(rgb: (f32, f32, f32)) -> Self {
+    pub fn from_tup(rgb: (f64, f64, f64)) -> Self {
         Self {
             x: rgb.0,
             y: rgb.1,
@@ -11,15 +11,15 @@ impl Color {
         }
     }
 
-    pub fn r(&self) -> f32 {
+    pub fn r(&self) -> f64 {
         self.x
     }
 
-    pub fn g(&self) -> f32 {
+    pub fn g(&self) -> f64 {
         self.y
     }
 
-    pub fn b(&self) -> f32 {
+    pub fn b(&self) -> f64 {
         self.z
     }
 }
