+++ src/ray.rs
@@ -14,7 +14,7 @@ impl Ray {
         }
     }
 
-    pub fn at(&self, t: f32) -> Vec3 {
+    pub fn at(&self, t: f64) -> Vec3 {
         self.base() + self.direction() * t
     }
